requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-cloud-aiplatform
google-cloud-secret-manager==2.17.0
python-gitlab==4.1.1
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
google-auth
vertexai
jinja2==3.1.2